requirements.txt
gensim==3.8.1
pyLDAvis==2.1.2
spacy==2.2.3
spacy.cli
en_core_web_lg
scikit-learn==0.22.2
seaborn==0.9.0
squarify==0.4.3
ipykernel
nltk
scipy
beautifulsoup4
Flask~=1.1.2
gunicorn~=20.0.4
psycopg2-binary~=2.8.5
flask_sqlalchemy~=2.4.3
flask_migrate~=2.5.3
pandas~=1.0.3
requests~=2.24.0
python-dotenv~=0.13.0